+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 7.2.0
+
+Merged [Pull Request](https://github.com/DarqueWarrior/vsteam/pull/371) from [Sebastian Schütze](https://github.com/SebastianSchuetze) which included the following:
+
+- Added Set-VSTeamPipelineBilling to buy or release Microsoft-hosted and self-hosted agents
+- Added Get-VSTeamAccounts to get the organizations where the user has access. Where the given user is either a member or an owner
+- Added Get-VSTeamUserProfile that gets the users profile of an account.
+- fixed filenames of files to work on linux (casing)
+
 ## 7.1.4
 
 Combined all the build json files (_types.json, _formats.json and _functions.json) into a single file called config.json in the root folder.